Add files via upload
A soil moisture  sensor I developed is an essential component in embedded systems designed for monitoring and controlling soil moisture levels. It is used to measure the amount of water present in the soil, providing valuable information for irrigation management, plant health monitoring, and environmental analysis.

The sensor typically consists of a probe or electrodes that are inserted into the soil. These electrodes make direct contact with the soil, allowing them to measure the electrical conductivity or impedance of the soil, which correlates with its moisture content. The sensor may employ different technologies such as resistive, capacitive, or frequency-based methods to measure soil moisture.

In an embedded system, the soil moisture sensor is connected to a microcontroller or microprocessor that interfaces with the sensor and processes the data.